Remove usage of qDebug() in extract callback.
